// apps/web/src/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)";
  };
  graphql_public: {
    Tables: {
      [_ in never]: never;
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      graphql: {
        Args: {
          operationName?: string;
          query?: string;
          variables?: Json;
          extensions?: Json;
        };
        Returns: Json;
      };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
  public: {
    Tables: {
      profiles: {
        Row: {
          avatar_url: string | null;
          created_at: string;
          email: string | null;
          first_name: string | null;
          id: string;
          last_name: string | null;
          org: string | null;
          title: Database["public"]["Enums"]["Titles"] | null;
        };
        Insert: {
          avatar_url?: string | null;
          created_at?: string;
          email?: string | null;
          first_name?: string | null;
          id?: string;
          last_name?: string | null;
          org?: string | null;
          title?: Database["public"]["Enums"]["Titles"] | null;
        };
        Update: {
          avatar_url?: string | null;
          created_at?: string;
          email?: string | null;
          first_name?: string | null;
          id?: string;
          last_name?: string | null;
          org?: string | null;
          title?: Database["public"]["Enums"]["Titles"] | null;
        };
        Relationships: [];
      };
      projects: {
        Row: {
          basemap: Database["public"]["Enums"]["Basemaps"] | null;
          created_at: string | null;
          description: string | null;
          id: string;
          layers: Database["public"]["Enums"]["Layers"][] | null;
          preview_url: string | null;
          title: string | null;
          updated_at: string | null;
          user_id: string;
        };
        Insert: {
          basemap?: Database["public"]["Enums"]["Basemaps"] | null;
          created_at?: string | null;
          description?: string | null;
          id?: string;
          layers?: Database["public"]["Enums"]["Layers"][] | null;
          preview_url?: string | null;
          title?: string | null;
          updated_at?: string | null;
          user_id?: string;
        };
        Update: {
          basemap?: Database["public"]["Enums"]["Basemaps"] | null;
          created_at?: string | null;
          description?: string | null;
          id?: string;
          layers?: Database["public"]["Enums"]["Layers"][] | null;
          preview_url?: string | null;
          title?: string | null;
          updated_at?: string | null;
          user_id?: string;
        };
        Relationships: [];
      };
      trash: {
        Row: {
          basemap: Database["public"]["Enums"]["Basemaps"] | null;
          created_at: string | null;
          description: string | null;
          id: string;
          layers: Database["public"]["Enums"]["Layers"][] | null;
          preview_url: string | null;
          title: string | null;
          trashed_at: string | null;
          updated_at: string | null;
          user_id: string;
        };
        Insert: {
          basemap?: Database["public"]["Enums"]["Basemaps"] | null;
          created_at?: string | null;
          description?: string | null;
          id?: string;
          layers?: Database["public"]["Enums"]["Layers"][] | null;
          preview_url?: string | null;
          title?: string | null;
          trashed_at?: string | null;
          updated_at?: string | null;
          user_id?: string;
        };
        Update: {
          basemap?: Database["public"]["Enums"]["Basemaps"] | null;
          created_at?: string | null;
          description?: string | null;
          id?: string;
          layers?: Database["public"]["Enums"]["Layers"][] | null;
          preview_url?: string | null;
          title?: string | null;
          trashed_at?: string | null;
          updated_at?: string | null;
          user_id?: string;
        };
        Relationships: [];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      Basemaps: "Satellite" | "Light" | "Dark" | "Streets" | "Topographic";
      Layers: "Parcels";
      Titles:
        | "Analyste immobilier"
        | "Architecte"
        | "Courtier immobilier"
        | "Économiste"
        | "Évaluateur agréé"
        | "Investisseur Immobilier"
        | "Promoteur immobilier"
        | "Urbaniste"
        | "Autre";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never;

export const Constants = {
  graphql_public: {
    Enums: {},
  },
  public: {
    Enums: {
      Basemaps: ["Satellite", "Light", "Dark", "Streets", "Topographic"],
      Layers: ["Parcels"],
      Titles: [
        "Analyste immobilier",
        "Architecte",
        "Courtier immobilier",
        "Économiste",
        "Évaluateur agréé",
        "Investisseur Immobilier",
        "Promoteur immobilier",
        "Urbaniste",
        "Autre",
      ],
    },
  },
} as const;